requirements.txt
requests~=2.28.1
opcua~=0.98.13
cryptography~=41.0.1
pytz~=2020.5
psycopg2-binary~=2.9.3